add_ball now works with one of the compression strategies assumed in our paper. The amount of degrees of freedom in CDD2 is as huge as always. One reason could be that J in APPLY is overestimated. This could also be a reason for general CDD2 routines (also in 1 dim). Some testing needs to be done.

+++ WaveletTL/tests/test_tbasis_index.cpp
@@ -163,7 +163,7 @@ int main()
     
 #if 1
     // some output for debugging
-    for (unsigned int i=0; i<TBasisArray[0]->degrees_of_freedom(); ++i)
+    for (unsigned int i=0; (int)i<TBasisArray[0]->degrees_of_freedom(); ++i)
     {
         cout << "i = " << i << "; lambda(i) = " << *TBasisArray[0]->get_wavelet(i) << endl;
     }
